Move solvers to new module, add CVODE
