+++ src/thevenin/plotutils/__init__.py
@@ -1,6 +1,4 @@
 """
-Plotting Utilities
-------------------
 A module designed to enhance plotting with the matplotlib library. Helper
 functions include routines for simplifying color scheme management, formatting
 axis ticks, fonts, and more, making it easier to create polished and consistent
