+++ src/thevenin/solvers/_cvode.py
@@ -1,0 +1,9 @@
+from sksundae import cvode
+
+
+class CVODEResult(cvode.CVODEResult):
+    pass
+
+
+class CVODESolver(cvode.CVODE):
+    pass
