+++ src/thevenin/solvers/_ida.py
